Chuoyue-Algorithm-Server/requirements.txt
SQLAlchemy==1.3.19
opencv-python==3.4.2.16
numpy==1.19.2
pandas==1.1.3
Flask==1.1.2
torchvision==0.7.0
tqdm==4.49.0
matplotlib==3.3.1
torch==1.6.0
Pillow==7.2.0
tensorboardX==2.1
torchsummary==1.5.1
pymysql
cryptography